AMQP-426 Refactor MessagingMessageListnerAdapter
JIRA: https://jira.spring.io/browse/AMQP-426

Allow existing applications that are not using SF 4.1 features
to continue to use SF 4.0.

Essentially, push all spring-messaging dependencies down
to `MessagingMessageListenerAdapter`.

NOTE: Also skips conversion if the result is already a
o.s.a.c.Message.
